refactor(data-fetchers): move to dedicated directory and rename for clarity
- Move data-fetchers from controllers/data-fetchers/ to data-fetchers/
- Rename syntheticDataFetcher -> hundredMockDataFetcher (hundred-mock.ts)
- Rename mockDataFetcher -> thousandMockDataFetcher (thousand-mock.ts)
- Fix setupSemanticController overriding perf test mock fetcher
- Update all imports across Vue, React, and core
