+++ packages/core/src/data-fetchers/hundred-mock.ts
@@ -1,6 +1,6 @@
-import type { DataFetcher, KLineData } from '../types'
+import type { DataFetcher, KLineData } from '../controllers/types'
 
-export const syntheticDataFetcher: DataFetcher = async (_source, config) => {
+export const hundredMockDataFetcher: DataFetcher = async (_source, config) => {
   const start = new Date(config.startDate).getTime()
   const end = new Date(config.endDate).getTime()
   const dayMs = 86400000
