+++ packages/core/src/data-fetchers/index.ts
@@ -1,0 +1,3 @@
+export { thousandMockDataFetcher } from './thousand-mock'
+export { hundredMockDataFetcher } from './hundred-mock'
+export { baostockDataFetcher } from './baostock'
